divide pytest_clld into submodules, prevent direct imports
- add selenium.get_page()
- fix bare excepts
- refactor DataTable.get_info w/ Info.fromtext()
- unify time.sleep() calls
